Se crea API Rest de lugar, cliente y se agrega Basic Auth para proteger endpoints

+++ src/main/java/com/oo2/grupo17/entities/Contacto.java
@@ -29,7 +29,7 @@ public class Contacto {
 	private String email;
 	
 	@Column(name = "movil")
-	private int movil;
+	private long movil;
 	
 	@Column(name = "telefono")
 	private int telefono;
